backend/requirements.txt

fastapi==0.115.0
uvicorn[standard]==0.32.0
sqlalchemy==2.0.35
psycopg2-binary==2.9.10
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.10
email-validator==2.1.1
python-dotenv==1.0.1  # Optional, but safe